moved run_package_method and tests (#6988)
* moved run_package_method and tests

* fix tests

* remove error
