+++ conans/client/cmd/export_pkg.py
@@ -1,6 +1,7 @@
 import os
 
 from conans.client import packager
+from conans.client.conanfile.package import run_package_method
 from conans.client.graph.graph import BINARY_SKIP
 from conans.client.graph.graph_manager import load_deps_info
 from conans.errors import ConanException
@@ -61,9 +62,9 @@ def export_pkg(app, recorder, full_ref, source_folder, build_folder, package_fol
             prev = packager.export_pkg(conanfile, package_id, package_folder, dest_package_folder,
                                        hook_manager, conan_file_path, ref)
         else:
-            prev = packager.run_package_method(conanfile, package_id, source_folder, build_folder,
-                                               dest_package_folder, install_folder, hook_manager,
-                                               conan_file_path, ref, local=True)
+            prev = run_package_method(conanfile, package_id, source_folder, build_folder,
+                                      dest_package_folder, install_folder, hook_manager,
+                                      conan_file_path, ref, local=True)
 
     packager.update_package_metadata(prev, layout, package_id, full_ref.revision)
     pref = PackageReference(pref.ref, pref.id, prev)
