+++ conans/test/functional/conanfile/user_info_test.py
