+++ conans/test/unittests/client/packager_test.py
@@ -3,9 +3,9 @@ import shutil
 import unittest
 from mock import Mock
 
+from conans.client.conanfile.package import run_package_method
 from conans.client.graph.python_requires import ConanPythonRequire
 from conans.client.loader import ConanFileLoader
-from conans.client.packager import run_package_method
 from conans.model.ref import ConanFileReference, PackageReference
 from conans.paths import CONANFILE, CONANINFO
 from conans.test.utils.cpp_test_files import cpp_hello_source_files
